(BSR)[API] chore: Get dry about offerer address business logic
We are repeating same logic about offerer address/address creation when
creating/updating venue and/or offers.
